refactor: rename session/internal/shared.ts to utils.ts
