+++ packages/controller/src/session/index.ts
@@ -4,4 +4,4 @@ export * from "../errors";
 export * from "../types";
 export * from "./internal/types";
 export { CartridgeSessionAccount } from "./internal/account";
-export { signerToGuid } from "./internal/shared";
+export { signerToGuid } from "./internal/utils";
