+++ packages/controller/src/session/internal/index.ts
@@ -4,4 +4,4 @@ export * from "./merkle";
 export * from "./execution";
 export * from "./account";
 export * from "./subscribe";
-export * from "./shared";
+export * from "./utils";
